requirement.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase-py
google-generativeai
